Began Coding
I started laying out the project and working on it so we don't get behind. Feel free to edit your parts as necessary; it was just helpful to me to see it laid out and how it's interwoven as well as to make the subclasses without all the error messages. I'll probably work on it again before break is over. Through this, I discovered we are missing some things from our schedule and added them in blue. Please check the Google Sheet again https://docs.google.com/spreadsheets/d/1PNwDyBBMzftccGr-OzGQKD_mAge6H75SSlUnHq-eIZk/edit?usp=sharing